Added separated Provider
